+++ pkg/scheduler/actions/utils/input_jobs.go
@@ -44,6 +44,16 @@ func (jobsOrder *JobsOrderByQueues) InitializeWithJobs(
 			continue
 		}
 
+		// Skip jobs whose queue doesn't exist
+		if _, found := jobsOrder.ssn.Queues[job.Queue]; !found {
+			continue
+		}
+
+		// Skip jobs whose queue's parent queue doesn't exist
+		if _, found := jobsOrder.ssn.Queues[jobsOrder.ssn.Queues[job.Queue].ParentQueue]; !found {
+			continue
+		}
+
 		jobsOrder.addJobToQueue(job, jobsOrder.jobsOrderInitOptions.VictimQueue)
 	}